requirements.txt
mss>=9.0.0
Pillow>=10.0.0
pyautogui>=0.9.54
pywin32>=306
requests>=2.31.0
